+++ resources/volume.rb
@@ -6,7 +6,9 @@ state_attrs :access_key_id,
             :force,
             :device,
             :instance_id,
-            :retention_check
+            :retention_check,
+            :wait_for,
+            :timeout
 
 attribute :access_key_id,         :kind_of => String
 attribute :secret_access_key,     :kind_of => String
@@ -15,6 +17,8 @@ attribute :force,                 :kind_of => [TrueClass, FalseClass], :default 
 attribute :device,                :kind_of => String
 attribute :instance_id,           :kind_of => String
 attribute :retention_check,       :kind_of => [TrueClass, FalseClass], :default => false
+attribute :wait_for,              :kind_of => String
+attribute :timeout,               :kind_of => Integer
 
 def initialize(*args)
   super