+++ CHANGELOG.md
@@ -3,6 +3,16 @@ blockdevice-nativex CHANGELOG
 
 This file is used to list changes made in each version of the blockdevice-nativex cookbook.
 
+0.3.0
+-----
+- [Jesse Hauf] - Renamed snapshots to snapshots_take
+- [Jesse Hauf] - Created snapshots_restore recipe and added support for restoring snapshots
+- [Jesse Hauf] - Added lwrps to support independent create and attach operations, deleting volumes, and force
+detaching volumes.
+- [Jesse Hauf] - Added lwrp to wait for attach and create operations
+- [Jesse Hauf] - Add libraries to support aws-sdk-v1 Ruby operations
+- [Jesse Hauf] - Added timeout attribute support to volume recipe
+
 0.2.0
 -----
 - [Jesse Hauf] - Added support for HVM instances